initial extraction of Connection type
